Initial commit for styled individual profile cards.
