+++ src/components/lib/ProfileCard/index.scss
@@ -1,18 +1,12 @@
 @import "../../../styles/variables";
-
-.big-container {
-  margin: 2rem auto;
-  color: $black;
-  font-weight: 600;
-  background: $blue-100;
-  border-radius: 15px;
-  line-height: 2.5rem;
-  flex-direction: column;
-  padding: 2.5rem 1rem;
-  .temp-container {
-    width: 100%;
+.individual-cards {
+  display: flex;
+  .progress-card {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    align-items: center;
   }
-
   .profile-info {
     .profile-track {
       color: $blue-200;
@@ -21,33 +15,24 @@
   .toggle-assign {
     color: $pink-100;
   }
-}
 
-.req-title {
-  margin: 1rem;
-}
-.profile-card-container {
-  width: 25%;
-}
+  .req-title {
+    margin: 1rem;
+  }
+  .profile-card-container {
+    width: 25%;
+  }
+  .progress-number {
+    color: $black;
+  }
 
-.overall-progress {
-  width: 25%;
-}
-.progress-number {
-  color: $black;
-}
-
-.progress-card {
-  width: 33%;
-  align-items: center;
-}
-
-.req-container {
-  display: flex;
-}
-
-.progress-ring {
-  justify-content: center;
-  align-items: center;
-  margin: 1rem;
+  .progress-ring {
+    justify-content: center;
+    align-items: center;
+    margin: 1rem;
+    width: 150px;
+    .progress-number {
+      font-size: 3.2rem;
+    }
+  }
 }
