+++ src/components/pages/MyStudents/index.js
@@ -2,7 +2,7 @@ import React, { useEffect } from "react"
 import { useSelector, useDispatch } from "react-redux"
 import { fetchPinnedStudents } from "../../../actions"
 import { Link } from "react-router-dom"
-import ProfileCard from "../../lib/ProfileCard"
+import PinnedStudentCard from "../../lib/PinnedStudentCard"
 import FindStudent from "../../pages/FindStudent"
 
 import { ContentHeader, DashboardContent } from "../../layout"
@@ -22,25 +22,21 @@ const MyStudents = () => {
     dispatch(fetchPinnedStudents())
   }, [dispatch])
   return (
-    <>
-      <DashboardContent>
-        <ContentHeader
-          title={"Find a Student"}
-          prompt={"Search for a student by name"}
-          progress={-1}
-        ></ContentHeader>
-        <FindStudent />
-        <section className="my-pinned-students">
-          {pinnedStudents.map(student => (
-            <div className="student-profile-card" key={student.id}>
-              <Link to={`/student/${student.id}`}>
-                <ProfileCard student={student} showFull={"false"} />
-              </Link>
-            </div>
-          ))}
-        </section>
-      </DashboardContent>
-    </>
+    <DashboardContent>
+      <ContentHeader
+        title={"Find a Student"}
+        prompt={"Search for a student by name"}
+        progress={-1}
+      ></ContentHeader>
+      <FindStudent />
+      {pinnedStudents.map(student => (
+        <div className="student-profile-card" key={student.id}>
+          <Link to={`/student/${student.id}`}>
+            <PinnedStudentCard student={student}></PinnedStudentCard>
+          </Link>
+        </div>
+      ))}
+    </DashboardContent>
   )
 }
